fix(lab_result, paramter_range): changed upper and lower limit type to float

+++ core/entity/lab_result.py
@@ -10,8 +10,8 @@ class Lab_Result(BaseModel):
     parameter_id : Optional[int] = None
     unit : Optional[str] = None
     result : Optional[str] = None
-    upper_limit : Optional[int] = None
-    lower_limit : Optional[int] = None
+    upper_limit : Optional[float] = None
+    lower_limit : Optional[float] = None
     remark : Optional[str] = None
     created_time: Optional[datetime] = None
     updated_time: Optional[datetime] = None
